Remove jasmine-jquery, install more recent version of node in Travis
