+++ .travis.yml
@@ -20,12 +20,18 @@ before_install:
   - export TZ=America/New_York
   - echo psql --version
   - gem install -v 1.10.6 bundler --no-rdoc --no-ri
- # workarounds for phantomjs 1, see: https://github.com/travis-ci/travis-ci/issues/3225#issuecomment-177592725
+
+  # workarounds for phantomjs 1, see: https://github.com/travis-ci/travis-ci/issues/3225#issuecomment-177592725
   - mkdir travis-phantomjs
   - wget https://studentinsights-public.s3.amazonaws.com/phantomjs-2.1.1-linux-x86_64.tar.bz2 -O $PWD/travis-phantomjs/phantomjs-2.1.1-linux-x86_64.tar.bz2
   - tar -xvf $PWD/travis-phantomjs/phantomjs-2.1.1-linux-x86_64.tar.bz2 -C $PWD/travis-phantomjs
   - export PATH=$PWD/travis-phantomjs/phantomjs-2.1.1-linux-x86_64/bin:$PATH
-  - rm -rf ~/.nvm && git clone https://github.com/creationix/nvm.git ~/.nvm && (cd ~/.nvm && git checkout `git describe --abbrev=0 --tags`) && source ~/.nvm/nvm.sh && nvm install $TRAVIS_NODE_VERSION
+
+  # install node
+  - curl -sL https://deb.nodesource.com/setup_6.x | sudo -E bash - 
+  - sudo apt-get install -y nodejs
+  - node --version
+  - npm --version
   - npm install
 
 install:
@@ -45,7 +51,6 @@ env:
     # For AWS deploy, add in credentials for uploading to S3 and pushing to Docker Hub here:
     - secure: CVsj/OoauR/2rTan1fAriyW/b92DahJsQ3zh7MAHAeQZ3atRQq4cYxd6fRSO4zoAOeTTTHp4hQ3YX+Uo765bNXsUJLKiKT8Wnin+eOG8nMGxzmKM6oxuWIt552pbwvhuX3+GS8QDT1sbHAdedlDng/97BhXeZTuzgVXN6Zigdo8=
     - secure: jSWLU8N9ui5BZ7PmWmY08BrrOx1BN1IEdYhrTnv5RDWMbUyGTu62hyXOseKMy1b+URaWSiklkxLBK0uVm1oSYf/MYiDhNwFTqPn8eH7AK7MJcxSOaJ4ce0eCEeTy56Wnbljjt0jYeGMFS05uBGhUT82vYvXq4l4h7pCORJWXaPA=
-    - TRAVIS_NODE_VERSION="7"
 
 notifications:
   webhooks: http://project-monitor.codeforamerica.org/projects/somerville-teacher-tool/status
